Add LanguagePanel.
